refactor: move utils and models to pkg directory and update imports
