--- internal/models/google_user.go
@@ -1,12 +1,0 @@
-package models
-
-// GoogleUser represents the user information returned by Google's OAuth2 API
-type GoogleUser struct {
-	ID            string `json:"id"`
-	Email         string `json:"email"`
-	VerifiedEmail bool   `json:"verified_email"`
-	Name          string `json:"name"`
-	GivenName     string `json:"given_name"`
-	FamilyName    string `json:"family_name"`
-	Picture       string `json:"picture"`
-}
